requirements.txt
Package             Version
------------------- -------
asgiref             3.2.10
cffi                1.14.0
cryptography        2.9.2
Django              3.0.7
django-rest-knox    4.1.0
django-solo         1.1.3
djangorestframework 3.11.0
importlib-metadata  1.6.1
Markdown            3.2.2
Pillow              7.1.2
pip                 20.2.3
psycopg2            2.8.6
pycparser           2.20
python-decouple     3.3
pytz                2020.1
setuptools          49.6.0
six                 1.15.0
sqlparse            0.3.1
wheel               0.35.1
zipp                3.1.0
